+++ src/IEventSubscriber.php
@@ -1,0 +1,9 @@
+<?php
+declare(strict_types=1);
+
+namespace Konecnyjakub\EventDispatcher;
+
+interface IEventSubscriber
+{
+    public static function getSubscribedEvents(): iterable;
+}