+++ tests/ListenerProviderTest.php
@@ -25,5 +25,32 @@ final class ListenerProviderTest extends TestCase
         $listenerProvider->registerListeners(Event::class, ["time", "pi", ]);
         $this->assertSame(["time", "pi", ], $listenerProvider->getListenersForEvent(new Event()));
         $this->assertSame([], $listenerProvider->getListenersForEvent(new \stdClass()));
+
+        $eventSubscriber = new class implements IEventSubscriber
+        {
+            public function one(): void
+            {
+            }
+
+            public function two(): void
+            {
+            }
+
+            public static function getSubscribedEvents(): iterable
+            {
+                return [
+                    Event::class => [
+                        ["one", ], ["two", 1, ],
+                    ]
+                ];
+            }
+        };
+        $listenerProvider = new ListenerProvider();
+        $listenerProvider->addSubscriber($eventSubscriber);
+        $this->assertSame(
+            [[$eventSubscriber, "one"], [$eventSubscriber, "two"], ],
+            $listenerProvider->getListenersForEvent(new Event())
+        );
+        $this->assertSame([], $listenerProvider->getListenersForEvent(new \stdClass()));
     }
 }