+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Version 1.2.0-dev
 - allowed registering multiple listeners at the same time in PriorityListenerProvider
 - added default priority for listeners in PriorityListenerProvider
 - added DebugEventDispatcher
+- added support for event subscribers
 
 Version 1.1.0
 - allowed registering multiple listeners at the same time in ListenerProvider